AuthService/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
python-jose==3.3.0
passlib[argon2]==1.7.4
argon2-cffi==23.1.0
pydantic==1.10.13
python-dotenv==1.0.1
email-validator==2.1.0.post1
